pyproject.toml
[project]
name = "line-agui-adapter"
version = "0.1.0"
description = "AG-UI adapter for line-bot-sdk-python v3"
readme = "README.md"
authors = [
    { name = "Sogo Kato", email = "sogo.kato@outlook.com" }
]
requires-python = ">=3.13"
dependencies = [
    "ag-ui-protocol>=0.1.14",
    "httpx>=0.27.0",
    "line-bot-sdk>=3.0.0",
    "pydantic>=2.0.0",
]

[build-system]
requires = ["uv_build>=0.9.26,<0.10.0"]
build-backend = "uv_build"

[dependency-groups]
dev = [
    "fastapi>=0.135.1",
    "ruff>=0.15.7",
    "uvicorn>=0.42.0",
]

[tool.ruff]
line-length = 88

[tool.ruff.lint]
select = [
    "D",
    "E",
    "F",
    "I",
    "N",
    "W",
]
ignore = [
    "D203",
    "D213",
]
